requirements.txt
Flask
google-generativeai
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-core
faiss-cpu
python-dotenv
Werkzeug
Pillow
